Moved hooks to correct folder (#8419)
